improve and unify common logic between all token update and create system contracts
