+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/exec/systemcontracts/hts/burn/FungibleBurnCallTest.java
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.burn;
+package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.hts.burn;
 
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_TOKEN_BURN_AMOUNT;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_TOKEN_ID;
 import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnTranslator.BURN_TOKEN_V1;
 import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.burn.BurnTranslator.BURN_TOKEN_V2;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.A_NEW_ACCOUNT_ID;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.EIP_1014_ADDRESS;
-import static com.hedera.node.app.service.contract.impl.test.TestHelpers.NON_FUNGIBLE_TOKEN_ID;
+import static com.hedera.node.app.service.contract.impl.test.TestHelpers.FUNGIBLE_TOKEN_ID;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.asBytesResult;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.asHeadlongAddress;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -42,7 +43,7 @@ import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-public class NonFungibleBurnCallTest extends HtsCallTestBase {
+public class FungibleBurnCallTest extends HtsCallTestBase {
 
     private static final org.hyperledger.besu.datatypes.Address FRAME_SENDER_ADDRESS = EIP_1014_ADDRESS;
 
@@ -58,11 +59,21 @@ public class NonFungibleBurnCallTest extends HtsCallTestBase {
     private FungibleBurnCall subject;
 
     @Test
-    void happyPathV1() {
-        givensForSuccessfulNFTBurnV1AndV2();
-        given(recordBuilder.status()).willReturn(ResponseCodeEnum.SUCCESS);
+    void revertsOnMissingToken() {
+        subject = new FungibleBurnCall(
+                1234, mockEnhancement(), null, verificationStrategy, FRAME_SENDER_ADDRESS, addressIdConverter);
 
-        subject = subjectForNFTBurn(1L);
+        final var result = subject.execute().fullResult().result();
+
+        assertEquals(MessageFrame.State.REVERT, result.getState());
+        assertEquals(Bytes.wrap(INVALID_TOKEN_ID.protoName().getBytes()), result.getOutput());
+    }
+
+    @Test
+    void burnTokenHappyPathV1() {
+        givensForSuccessfulBurnTokenV1AndV2();
+
+        subject = subjectForBurn(10L);
 
         final var result = subject.execute().fullResult().result();
 
@@ -75,11 +86,10 @@ public class NonFungibleBurnCallTest extends HtsCallTestBase {
     }
 
     @Test
-    void happyPathV2() {
-        givensForSuccessfulNFTBurnV1AndV2();
-        given(recordBuilder.status()).willReturn(ResponseCodeEnum.SUCCESS);
+    void burnTokenHappyPathV2() {
+        givensForSuccessfulBurnTokenV1AndV2();
 
-        subject = subjectForNFTBurn(1L);
+        subject = subjectForBurn(10L);
 
         final var result = subject.execute().fullResult().result();
 
@@ -93,19 +103,6 @@ public class NonFungibleBurnCallTest extends HtsCallTestBase {
 
     @Test
     void unhappyPathRevertsWithReason() {
-        givensForSuccessfulNFTBurnV1AndV2();
-
-        given(recordBuilder.status()).willReturn(INVALID_TOKEN_BURN_AMOUNT);
-
-        subject = subjectForNFTBurn(1L);
-
-        final var result = subject.execute().fullResult().result();
-
-        assertEquals(MessageFrame.State.REVERT, result.getState());
-        assertEquals(Bytes.wrap(INVALID_TOKEN_BURN_AMOUNT.protoName().getBytes()), result.getOutput());
-    }
-
-    private void givensForSuccessfulNFTBurnV1AndV2() {
         given(addressIdConverter.convert(asHeadlongAddress(FRAME_SENDER_ADDRESS)))
                 .willReturn(A_NEW_ACCOUNT_ID);
         given(systemContractOperations.dispatch(
@@ -114,13 +111,33 @@ public class NonFungibleBurnCallTest extends HtsCallTestBase {
                         eq(A_NEW_ACCOUNT_ID),
                         eq(TokenBurnRecordBuilder.class)))
                 .willReturn(recordBuilder);
+        given(recordBuilder.status()).willReturn(INVALID_TOKEN_BURN_AMOUNT);
+
+        subject = subjectForBurn(1L);
+
+        final var result = subject.execute().fullResult().result();
+
+        assertEquals(MessageFrame.State.REVERT, result.getState());
+        assertEquals(Bytes.wrap(INVALID_TOKEN_BURN_AMOUNT.protoName().getBytes()), result.getOutput());
     }
 
-    private FungibleBurnCall subjectForNFTBurn(final long serialNo) {
+    private void givensForSuccessfulBurnTokenV1AndV2() {
+        given(addressIdConverter.convert(asHeadlongAddress(FRAME_SENDER_ADDRESS)))
+                .willReturn(A_NEW_ACCOUNT_ID);
+        given(systemContractOperations.dispatch(
+                        any(TransactionBody.class),
+                        eq(verificationStrategy),
+                        eq(A_NEW_ACCOUNT_ID),
+                        eq(TokenBurnRecordBuilder.class)))
+                .willReturn(recordBuilder);
+        given(recordBuilder.status()).willReturn(ResponseCodeEnum.SUCCESS);
+    }
+
+    private FungibleBurnCall subjectForBurn(final long amount) {
         return new FungibleBurnCall(
-                serialNo,
+                amount,
                 mockEnhancement(),
-                NON_FUNGIBLE_TOKEN_ID,
+                FUNGIBLE_TOKEN_ID,
                 verificationStrategy,
                 FRAME_SENDER_ADDRESS,
                 addressIdConverter);
